Fixed #1473: Show rank pts with precision 1
git-svn-id: https://localhost/svn/wxTTM/trunk@8907 7754d8ab-9b54-e011-a251-0001c006bdc7

+++ src/wxTTM/Dialogs/ImportOnlineEntries.cpp
@@ -58,7 +58,7 @@ static int GetInt(XmlRpcValue &val)
 }
 
 
-static int GetDouble(XmlRpcValue &val)
+static double GetDouble(XmlRpcValue &val)
 {
   switch (val.getType())
   {
